+++ src/main/java/com/iteo/shopping/products/ProductDto.java
@@ -3,5 +3,7 @@ package com.iteo.shopping.products;
 import org.springframework.lang.NonNull;
 import org.springframework.lang.Nullable;
 
-public record ProductDto(@NonNull String name, @Nullable Integer price) {
+import java.math.BigDecimal;
+
+public record ProductDto(@NonNull String name, @Nullable BigDecimal price) {
 }
