+++ src/main/java/com/iteo/shopping/products/ProductView.java
@@ -1,6 +1,7 @@
 package com.iteo.shopping.products;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record ProductView(UUID uuid, String name, Integer price) {
+public record ProductView(UUID uuid, String name, BigDecimal price) {
 }
